requirements.txt
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
